refactor(cli): split observer process startup internals
